+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+6.6.37
+- Billing: fix error in Document generator order when contact not entered
+
+28-01-2022 6.6.36
+- Tasks: drag and drop tasks to other lists
+- Tasks: show green dates when active, show red when late. removed bold when status is needs action.
+- Tasks: remember filter states
+- Tasks: Auto update due and start date if the other date changes
+- Tasks: Added due in 7 days filter
+- Tasks: Show categories in detail and grid
+
 27-01-2022 6.6.35
 - Tasks: Added widget for startpage
 - Tasks: When creating tasks from other Contacts, Projects or Orders the subject field will be auto filled.